Move JiraApiResource to Concerns dir
