+++ tests/Unit/ScrumMaster/Concerns/JiraApiResource.php
@@ -2,13 +2,16 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Unit\ScrumMaster;
+namespace App\Tests\Unit\ScrumMaster\Concerns;
 
 use App\ScrumMaster\Jira\Tickets;
 use PHPUnit\Framework\MockObject\MockObject;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 use Symfony\Contracts\HttpClient\ResponseInterface;
 
+/**
+ * @method createMock($originalClassName): MockObject
+ */
 trait JiraApiResource
 {
     private function mockJiraClient(array $issues): HttpClientInterface
